% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apeglmase.R
\name{apeglmase}
\alias{apeglmase}
\title{Helper function to fit apeglm model while handling common warnings/errors}
\usage{
apeglmase(y, total, mod.mat)
}
\arguments{
\item{y}{vector of counts to be modeled}

\item{total}{vector of total counts}

\item{mod.mat}{model matrix that contains the design matrix of
interest.}
}
\value{
A list the result, warnings, and errors
}
\description{
This function fits a beta-binomial model with shrinkage
from the package apeglm
and wraps it so that warnings and/or errors are output.
}
\details{
Just wraps apeglm to return either the fit or warning description.
}
