% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinase.R
\name{betabinase}
\alias{betabinase}
\title{Helper function to fit betabin model while handling common warnings/errors}
\usage{
betabinase(y, total, covariates = NULL)
}
\arguments{
\item{y}{vector of counts to be modeled}

\item{total}{vector of total counts}
}
\value{
A list the result, warnings, and errors
}
\description{
This function fits a beta-binomial model from the package aod
and wraps it so that warnings and/or errors are output.
}
\details{
Just wraps betabin to return either the fit or warning description. Warnings
and errors occur when the model doesn't converge. These will be flagged for
removing downstream.
}
