% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE_class.R
\name{build.designmatrix.intercept}
\alias{build.designmatrix.intercept}
\title{Constructs a design matrix for running CSIDE with only an intercept term}
\usage{
build.designmatrix.intercept(myRCTD, barcodes = NULL)
}
\arguments{
\item{myRCTD}{an \code{\linkS4class{RCTD}} object with annotated cell types e.g. from the \code{\link{run.RCTD}} function.}

\item{barcodes}{(default NULL) the barcodes, or pixel names, of the \code{\linkS4class{SpatialRNA}} object to be used when creating the design matrix.}
}
\value{
A design matrix containing the covariates for running CSIDE. The rownames represent pixel names and
are a subset of the pixels in the \code{\linkS4class{SpatialRNA}} object. The column represents the intercept
}
\description{
Constructs a design matrix for running CSIDE with only an intercept term
}
