% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cside_spase.R
\name{cside_spase}
\alias{cside_spase}
\title{Main function for detecting spatial allele-specific expression (ASE) patterns
considering cell type mixtures.}
\usage{
cside_spase(myRCTD, df = 5, genes = NULL, cores = 1, verbose = F)
}
\arguments{
\item{myRCTD}{RCTD object with maternalCounts and paternalCounts slots}

\item{df}{integer, sets the number of degrees of freedom to use for the
smoothing spline. Default is 5. Usually want to increase this or test multiple.}

\item{genes}{optional vector of genes to run on. Good for testing things out.}

\item{cores}{number of cores to use for parallelization. Default is 1.}
}
\value{
If method is betabinomial or quasibinomial, a list containing the
following output:
\itemize{ \item{\code{results}}{ a data frame containing a summary of
the results including p-values and q-values for goodness of spatial fit over
baseline covariates provided }
\item{\code{fits}}{ a list of beta-binomial model fit objects for each gene}}.
}
\description{
This function fits the spatial beta-binomial model with
cell type mixtures.
}
\details{
Fits the intercept-only model (only cell type effects) and the spatial
model (non-parametric thin plate smoothing splines)
}
