% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplineMatrix.R
\name{getSplineMatrix}
\alias{getSplineMatrix}
\title{Helper function to get normalized spline basis function matrix}
\usage{
getSplineMatrix(coords, df = 5)
}
\arguments{
\item{coords}{data frame or matrix of coordinates (either 1D or 2D)}

\item{df}{degrees of freedom to use for constructing the thin plate
regression spline. Default is 5.}
}
\value{
normalized spline basis function matrix
}
\description{
Takes in coordinates and degrees of freedom generates
the normalized spline basis function matrix using the `mgcv` `smoothCon`
function.
}
