% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScase.R
\name{plotScase}
\alias{plotScase}
\title{Main function for plotting ASE fits in single-cell data}
\usage{
plotScase(
  matrix1,
  matrix2,
  scasefit,
  genes = NULL,
  save = NULL,
  breaks = c(1, 2, 5, 10, 50, 100, 250, 500, 1000, 2000)
)
}
\arguments{
\item{matrix1}{a matrix of counts where  the rows are genes and the columns
are cells for allele 1. This one is the one that gets its probability
modeled. Must  have  row  names and column names to identify genes and cells.}

\item{matrix2}{a matrix of counts where the rows are genes and the columns
are cells for allele 2.}

\item{scasefit}{the resulting data frame from a call to scase}

\item{genes}{a vector of genes given as strings to plot. default is NULL
which will plot the first gene in the scasefit results dataframe.}

\item{save}{the filename prefix to save if want to save instead of visualize.
Default is null meaning no file will be saved.}

\item{breaks}{how to split up the total UMI counts for coloring. there is
a default setting which should work for most data sets.}
}
\value{
Multiple ggplot2 plots or, instead, each one gets saved to a different
file starting with the prefix defined in the save parameter.
}
\description{
This function plots stuff for scase results.
}
\details{
Takes in the result from scase and plots desired genes with beta distribution
overlaid on the histogram of allele1 proportion, confidence interval plot
for the desired genes, and the overall distribution of p and phi for all
genes
}
