% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasibinase.R
\name{quasibinase}
\alias{quasibinase}
\title{Helper function to fit quasibinomial model while handling common warnings/errors}
\usage{
quasibinase(y, total, mod.mat = NULL)
}
\arguments{
\item{y}{vector of counts to be modeled}

\item{total}{vector of total counts}

\item{mod.mat}{optional model matrix that contains the design matrix of
interest. Used for adding the thin plate regression spline basis functions
in the `spase` function and for adding cell type as a covariate. Note that
we assume there is an intercept column in mod.mat.}
}
\value{
A list the result, warnings, and errors
}
\description{
This function fits a quasibinomial model from the package aod
and wraps it so that warnings and/or errors are output.
}
\details{
Just wraps glm to return either the fit or warning description. Warnings
and errors occur when the model doesn't converge. These will be flagged for
removing downstream.
}
