% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIDE.R
\name{run.CSIDE.intercept}
\alias{run.CSIDE.intercept}
\title{Runs CSIDE in intercept-only mode}
\usage{
run.CSIDE.intercept(
  myRCTD,
  barcodes = NULL,
  cell_types = NULL,
  cell_type_threshold = 125,
  gene_threshold = 5e-05,
  doublet_mode = T,
  weight_threshold = NULL,
  sigma_gene = T,
  PRECISION.THRESHOLD = 0.05,
  cell_types_present = NULL,
  normalize_expr = F,
  logs = F,
  test_error = F,
  spase = F
)
}
\description{
Runs CSIDE in intercept-only mode
}
