% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scase.R
\name{scase}
\alias{scase}
\title{Main function for estimating ASE in single-cell data}
\usage{
scase(
  matrix1,
  matrix2,
  covariates = NULL,
  min.cells = 10,
  cores = 1,
  genes = NULL,
  add.var = 0,
  verbose = F
)
}
\arguments{
\item{matrix1}{a matrix of counts where  the rows are genes and the columns
are cells for allele 1. This one is the one that gets its probability
modeled. Must  have  row  names and column names to identify genes and cells.}

\item{matrix2}{a matrix of counts where the rows are genes and the columns
are cells for allele 2.}

\item{min.cells}{numeric specifying the minimum number of cells a gene
should be present on to fit. Default is 10 cells.}

\item{cores}{number of cores to use for parallelization. Default is 1.}

\item{genes}{which genes to fit the model on, should be  a  subset of  the
rownames of matrix1, default is all of the genes}

\item{add.var}{amount of additional variance  to add, default is 0. If
non-zero, addds additional columns to result data  frame suffixed with  .adj
to indicate calculated with the adjusted standard error}

\item{verbose}{whether or not to print a lot of status messages. Default is
FALSE.}
}
\value{
A data frame of results containing the gene names, estimated p and
standard errors on the logit scale.
}
\description{
This function fits a beta-binomial model and  returns estimates
of p and confidence intervals
}
\details{
Fits a beta-binomial model for each gene and returns a results data frame.
}
