% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spase.R
\name{spase}
\alias{spase}
\title{Main function for detecting spatial allele-specific expression (ASE) patterns}
\usage{
spase(
  se,
  predict = NULL,
  model = "all",
  min.pixels = 100,
  min.k = 100,
  max.k = 300,
  cores = 1,
  verbose = F
)
}
\arguments{
\item{predict}{(optional) a \code{data.frame} or \code{matrix} containing the
coordinates at which to evaluate the model. If \code{NULL} then a default
set of 400 coordinates is automatically generated from the range of the data.}

\item{model}{a string specifying which model to do, either 'all', 'cell.type',
or 'cell.type.interaction'. 'all' is the default and just uses all data to fit
the spatial model. 'cell.type' includes cell type as a covariate; each cell
type gets its own mean (intercept), 'cell.type.interaction' includes
interaction terms between cell type and the spatial spline, which allows each
cell type to have its own intercept and spatial fit. Typically challenging
to fit the latter two models with very sparse data.}

\item{min.pixels}{numeric specifying the minimum number of pixels a gene
should be present on to fit}

\item{min.k}{numeric, sets the minimum $k$ value for fitting the GAM model.
Default is \code{100}.}

\item{max.k}{numeric, sets the maximum $k$ value for fitting the GAM model.
Default is \code{300}.}

\item{cores}{number of cores to use for parallelization. Default is 1.}

\item{verbose}{whether or not to print a lot of status messages. Default is
FALSE.}

\item{data}{a data object containing gene expression counts for both alleles
(types). Accepted format is a
 \code{SpatialExperiment} object with exactly two assays, one for each
 allele. The assays can be sparse matrices and should be named by allele/type.
 By default, the probability of the first assay will be modeled as the
 binomial proportion p. Rows should be genes and columns should be pixel IDs.
 spatialCoords should optionally contain the cell types corresponding to each
 pixel.}
}
\value{
A list containing following output:
\itemize{ \item{\code{summaryStats}}{ a data frame containing a summary of
the results including approximate goodness of fit (chi-squared) p-values and
effect sizes of cell types if applicable}
\item{\code{fittedVals}}{ a matrix containing the fitted values for each
gene} \item{\code{seVals}}{ a matrix containing the standard errors of the
fitted values for each gene}}
}
\description{
This function fits a spatial binomial model, estimates a
smoothed probability surface, and stores information for plotting results.
Input must include either a list of two gene x pixel count data matrices or
a data frame containing a sparse representation of that, and a data frame or
matrix of pixel IDs and spatial coordinates. Optionally, single cell types
may be included for each pixel.
}
\details{
Fits a binomial Gaussian process GAM model for each gene,
extracts approximate goodness-of-fit (chi-squared) test p-values, and stores
predicted estimates and standard errors for visualization.
}
